Feat: POS Print

+++ database/migrations/2019_07_09_124359_create_sales_table.php
@@ -18,7 +18,7 @@ class CreateSalesTable extends Migration
             $table->unsignedInteger('sale_worker_id');
             $table->unsignedInteger('sale_client_id');
             $table->unsignedInteger('sale_payment_form_id');
-            $table->integer('sale_total_amount');
+            $table->decimal('sale_total_amount',8,2);
             $table->timestamps();
 
             $table->foreign('sale_worker_id')->references('id')->on('users');
